test: splitting general function to library class and use it as second super class
